house keeping
- Cleanup some eslint rules
- Lint some various jsons
- Convert all `.mjs` files back to standard `.js` files for less PM2 headaches in the long run
- Keep `ReactMap.mjs` as a polyfill with a deprecation warning after version 1.8.0
- Set all Location icons to the crosshairs style
- Cleanup old vscode extensions
- Format on save
