+++ .eslintrc
@@ -19,6 +19,7 @@
     "inject": true
   },
   "rules": {
+    "global-require": 0,
     "camelcase": 0,
     "react/jsx-props-no-spreading": 0,
     "linebreak-style": 0,
