+++ .prettierignore
@@ -3,3 +3,4 @@ node_modules
 public/missing-locales
 public/images/custom
 public/images/uicons
+server/src/configs/
