+++ .vscode/extensions.json
@@ -4,8 +4,7 @@
     "dbaeumer.vscode-eslint",
     "lokalise.i18n-ally",
     "esbenp.prettier-vscode",
-    "leizongmin.node-module-intellisense",
-    "eg2.vscode-npm-script",
+    "christian-kohler.npm-intellisense",
     "graphql.vscode-graphql"
   ]
 }
