+++ .vscode/settings.json
@@ -1,7 +1,5 @@
 {
-  "i18n-ally.localesPaths": [
-    "public/base-locales"
-  ],
+  "i18n-ally.localesPaths": ["public/base-locales"],
   "i18n-ally.keystyle": "flat",
   "[javascript]": {
     "editor.autoClosingBrackets": "always",
@@ -11,4 +9,5 @@
     "editor.autoClosingBrackets": "always",
     "editor.defaultFormatter": "esbenp.prettier-vscode"
   },
+  "editor.formatOnSave": true
 }
