+++ jsconfig.json
@@ -1,15 +1,15 @@
 {
   "compilerOptions": {
     "target": "es2017",
+    "module": "commonjs",
     "allowSyntheticDefaultImports": false,
     "baseUrl": "./src",
     "paths": {
       "@assets/*": ["./assets/*"],
       "@components/*": ["./components/*"],
       "@services/*": ["./services/*"],
-      "@hooks/*": ["./hooks/*"],
-      "@classes/*": ["./classes/*"],
+      "@hooks/*": ["./hooks/*"]
     }
   },
-  "exclude": ["node_modules", "dist"],
+  "exclude": ["node_modules", "**/node_modules/*", "dist"]
 }
