+++ package.json
@@ -1,17 +1,17 @@
 {
   "name": "reactmap",
-  "version": "1.5.0",
+  "version": "1.5.1",
   "description": "React based frontend map.",
-  "main": "ReactMap.mjs",
+  "main": "ReactMap.js",
   "author": "TurtIeSocks <58572875+TurtIeSocks@users.noreply.github.com>",
   "license": "MIT",
   "private": true,
   "scripts": {
     "start": "node .",
-    "build": "yarn create-locales && node esbuild.config.mjs",
+    "build": "yarn create-locales && node esbuild.config.js",
     "server": "yarn create-area && yarn generate && yarn migrate:latest && node server/src/index.js",
-    "watch": "node esbuild.config.mjs --dev",
-    "serve": "node esbuild.config.mjs --dev --serve",
+    "watch": "node esbuild.config.js --dev",
+    "serve": "node esbuild.config.js --dev --serve",
     "dev": "nodemon server/src/index.js --watch server",
     "generate": "node server/scripts/generateMasterfile.js",
     "gen-env-config": "node server/scripts/genEnvConfig.js",
