+++ server/src/data/defaultRarity.json
@@ -1,21 +1,58 @@
 {
   "common": [
-    1,4,7,10,13,16,19,21,23,25,27,29,32,35,37,39,41,43,46,48,50,52,54,56,58,60,63,66,69,72,74,77,79,81,84,86,88,90,92,95,96,98,100,102,104,109,111,116,118,120,123,129,133,138,140,152,155,158,161,163,165,167,170,177,179,183,187,190,191,193,194,198,200,204,207,209,215,216,218,220,223,228,231,252,255,258,261,263,265,270,273,276,278,280,283,285,287,293,296,299,300,304,307,309,311,312,315,316,318,320,322,325,328,331,333,339,341,343,345,347,349,351,353,355,361,363,366,370,387,390,393,396,399,401,412,415,418,420,421,425,427,431,434,436,449,451,453,456,459,495,498,501,504,506,509,517,519,527,529,535,540,543,546,548,551,557,562,568,570,572,577,580,582,585,588,590,592,602,605,613,616,619,624,629,650,653,656,659,661,667
+    1, 4, 7, 10, 13, 16, 19, 21, 23, 25, 27, 29, 32, 35, 37, 39, 41, 43, 46, 48,
+    50, 52, 54, 56, 58, 60, 63, 66, 69, 72, 74, 77, 79, 81, 84, 86, 88, 90, 92,
+    95, 96, 98, 100, 102, 104, 109, 111, 116, 118, 120, 123, 129, 133, 138, 140,
+    152, 155, 158, 161, 163, 165, 167, 170, 177, 179, 183, 187, 190, 191, 193,
+    194, 198, 200, 204, 207, 209, 215, 216, 218, 220, 223, 228, 231, 252, 255,
+    258, 261, 263, 265, 270, 273, 276, 278, 280, 283, 285, 287, 293, 296, 299,
+    300, 304, 307, 309, 311, 312, 315, 316, 318, 320, 322, 325, 328, 331, 333,
+    339, 341, 343, 345, 347, 349, 351, 353, 355, 361, 363, 366, 370, 387, 390,
+    393, 396, 399, 401, 412, 415, 418, 420, 421, 425, 427, 431, 434, 436, 449,
+    451, 453, 456, 459, 495, 498, 501, 504, 506, 509, 517, 519, 527, 529, 535,
+    540, 543, 546, 548, 551, 557, 562, 568, 570, 572, 577, 580, 582, 585, 588,
+    590, 592, 602, 605, 613, 616, 619, 624, 629, 650, 653, 656, 659, 661, 667
   ],
   "uncommon": [
-    2,5,8,11,14,17,20,22,24,26,30,33,36,38,40,42,44,47,49,51,53,55,57,59,61,64,67,70,73,75,78,80,82,85,87,89,91,93,97,99,101,103,105,110,112,117,119,121,124,125,126,127,139,141,153,156,159,162,164,166,168,171,178,180,184,185,188,195,202,203,205,206,210,211,213,217,219,221,224,226,227,229,234,253,256,259,262,264,266,268,271,274,277,279,281,284,286,288,294,297,301,302,305,308,310,317,319,323,326,329,332,340,342,344,346,348,354,356,362,364,388,391,394,397,400,402,419,426,428,432,435,437,450,452,454,457,460,496,499,502,505,507,510,520,522,524,528,536,541,544,547,549,552,558,563,569,573,578,581,583,586,587,591,595,603,606,614,615,620,651,654,657,660,662,668
+    2, 5, 8, 11, 14, 17, 20, 22, 24, 26, 30, 33, 36, 38, 40, 42, 44, 47, 49, 51,
+    53, 55, 57, 59, 61, 64, 67, 70, 73, 75, 78, 80, 82, 85, 87, 89, 91, 93, 97,
+    99, 101, 103, 105, 110, 112, 117, 119, 121, 124, 125, 126, 127, 139, 141,
+    153, 156, 159, 162, 164, 166, 168, 171, 178, 180, 184, 185, 188, 195, 202,
+    203, 205, 206, 210, 211, 213, 217, 219, 221, 224, 226, 227, 229, 234, 253,
+    256, 259, 262, 264, 266, 268, 271, 274, 277, 279, 281, 284, 286, 288, 294,
+    297, 301, 302, 305, 308, 310, 317, 319, 323, 326, 329, 332, 340, 342, 344,
+    346, 348, 354, 356, 362, 364, 388, 391, 394, 397, 400, 402, 419, 426, 428,
+    432, 435, 437, 450, 452, 454, 457, 460, 496, 499, 502, 505, 507, 510, 520,
+    522, 524, 528, 536, 541, 544, 547, 549, 552, 558, 563, 569, 573, 578, 581,
+    583, 586, 587, 591, 595, 603, 606, 614, 615, 620, 651, 654, 657, 660, 662,
+    668
   ],
   "rare": [
-    3,6,9,12,15,18,28,31,34,45,62,65,68,71,76,94,106,107,108,113,114,130,131,132,134,135,136,137,142,143,144,154,157,160,169,176,181,189,232,235,237,241,242,254,257,260,267,269,272,275,282,289,295,306,330,358,365,389,392,395,398,408,409,410,411,413,414,497,500,503,508,518,521,523,525,526,530,531,537,542,545,553,554,555,564,565,566,567,571,574,575,576,579,584,589,593,594,596,597,598,604,617,618,621,622,623,625,630,636,637,652,655,658
+    3, 6, 9, 12, 15, 18, 28, 31, 34, 45, 62, 65, 68, 71, 76, 94, 106, 107, 108,
+    113, 114, 130, 131, 132, 134, 135, 136, 137, 142, 143, 144, 154, 157, 160,
+    169, 176, 181, 189, 232, 235, 237, 241, 242, 254, 257, 260, 267, 269, 272,
+    275, 282, 289, 295, 306, 330, 358, 365, 389, 392, 395, 398, 408, 409, 410,
+    411, 413, 414, 497, 500, 503, 508, 518, 521, 523, 525, 526, 530, 531, 537,
+    542, 545, 553, 554, 555, 564, 565, 566, 567, 571, 574, 575, 576, 579, 584,
+    589, 593, 594, 596, 597, 598, 604, 617, 618, 621, 622, 623, 625, 630, 636,
+    637, 652, 655, 658
   ],
   "ultraRare": [
-    147,148,149,201,246,247,248,371,372,373,374,375,376,443,444,445,607,608,609,610,611,612,633,634,635,714,715
+    147, 148, 149, 201, 246, 247, 248, 371, 372, 373, 374, 375, 376, 443, 444,
+    445, 607, 608, 609, 610, 611, 612, 633, 634, 635, 714, 715
   ],
   "regional": [
-    83,115,122,128,214,222,313,314,324,335,336,337,338,357,369,417,422,423,439,441,455,480,481,482,511,512,513,514,515,516,538,539,550,556,561,626,631,632,707
+    83, 115, 122, 128, 214, 222, 313, 314, 324, 335, 336, 337, 338, 357, 369,
+    417, 422, 423, 439, 441, 455, 480, 481, 482, 511, 512, 513, 514, 515, 516,
+    538, 539, 550, 556, 561, 626, 631, 632, 707
   ],
   "never": [
-    172,173,174,175,182,186,192,196,197,199,208,212,225,230,233,236,238,239,240,290,291,292,298,303,321,327,334,350,352,359,360,367,368,403,404,405,406,407,416,424,429,430,433,438,439,440,442,446,447,448,458,461,462,463,464,465,466,467,468,469,470,471,472,473,474,475,476,477,478,479,532,533,534,559,560,599,600,601,627,628,677,678
+    172, 173, 174, 175, 182, 186, 192, 196, 197, 199, 208, 212, 225, 230, 233,
+    236, 238, 239, 240, 290, 291, 292, 298, 303, 321, 327, 334, 350, 352, 359,
+    360, 367, 368, 403, 404, 405, 406, 407, 416, 424, 429, 430, 433, 438, 439,
+    440, 442, 446, 447, 448, 458, 461, 462, 463, 464, 465, 466, 467, 468, 469,
+    470, 471, 472, 473, 474, 475, 476, 477, 478, 479, 532, 533, 534, 559, 560,
+    599, 600, 601, 627, 628, 677, 678
   ],
   "event": []
 }
