+++ server/src/routes/authRouter.js
@@ -1,4 +1,3 @@
-/* eslint-disable global-require */
 /* eslint-disable import/no-dynamic-require */
 /* eslint-disable no-console */
 const router = require('express').Router()
