+++ server/src/services/DiscordClient.js
@@ -2,7 +2,6 @@
 /* eslint-disable class-methods-use-this */
 /* eslint-disable no-await-in-loop */
 /* eslint-disable no-restricted-syntax */
-/* eslint-disable global-require */
 /* eslint-disable import/no-dynamic-require */
 /* global BigInt */
 const fs = require('fs')
