+++ server/src/services/config.js
@@ -1,5 +1,4 @@
 /* eslint-disable import/no-dynamic-require */
-/* eslint-disable global-require */
 /* eslint-disable no-console */
 process.env.NODE_CONFIG_DIR = `${__dirname}/../configs`
