+++ src/components/layout/dialogs/tutorial/data.json
@@ -93,121 +93,46 @@
       "xlKarp": true,
       "xsRat": true,
       "ivOr": {
-        "great": [
-          1,
-          100
-        ],
-        "ultra": [
-          1,
-          100
-        ],
-        "iv": [
-          0,
-          100
-        ],
-        "level": [
-          1,
-          35
-        ]
+        "great": [1, 100],
+        "ultra": [1, 100],
+        "iv": [0, 100],
+        "level": [1, 35]
       },
       "standard": {
         "enabled": false,
         "size": "md",
-        "iv": [
-          0,
-          100
-        ],
-        "atk_iv": [
-          0,
-          15
-        ],
-        "def_iv": [
-          0,
-          15
-        ],
-        "sta_iv": [
-          0,
-          15
-        ],
-        "level": [
-          1,
-          35
-        ],
-        "great": [
-          1,
-          100
-        ],
-        "ultra": [
-          1,
-          100
-        ]
+        "iv": [0, 100],
+        "atk_iv": [0, 15],
+        "def_iv": [0, 15],
+        "sta_iv": [0, 15],
+        "level": [1, 35],
+        "great": [1, 100],
+        "ultra": [1, 100]
       },
       "filter": {
         "1-163": {
           "enabled": false,
           "size": "md",
           "adv": "",
-          "iv": [
-            0,
-            100
-          ],
-          "atk_iv": [
-            0,
-            15
-          ],
-          "def_iv": [
-            0,
-            15
-          ],
-          "sta_iv": [
-            0,
-            15
-          ],
-          "level": [
-            1,
-            35
-          ],
-          "great": [
-            1,
-            100
-          ],
-          "ultra": [
-            1,
-            100
-          ]
+          "iv": [0, 100],
+          "atk_iv": [0, 15],
+          "def_iv": [0, 15],
+          "sta_iv": [0, 15],
+          "level": [1, 35],
+          "great": [1, 100],
+          "ultra": [1, 100]
         },
         "1-897": {
           "enabled": false,
           "size": "md",
           "adv": "",
-          "iv": [
-            0,
-            100
-          ],
-          "atk_iv": [
-            0,
-            15
-          ],
-          "def_iv": [
-            0,
-            15
-          ],
-          "sta_iv": [
-            0,
-            15
-          ],
-          "level": [
-            1,
-            35
-          ],
-          "great": [
-            1,
-            100
-          ],
-          "ultra": [
-            1,
-            100
-          ]
+          "iv": [0, 100],
+          "atk_iv": [0, 15],
+          "def_iv": [0, 15],
+          "sta_iv": [0, 15],
+          "level": [1, 35],
+          "great": [1, 100],
+          "ultra": [1, 100]
         }
       }
     },
